+++ module-services/service-desktop/endpoints/update/UpdateEndpoint.cpp
@@ -20,15 +20,22 @@ auto UpdateEndpoint::handle(Context &context) -> void
 auto UpdateEndpoint::run(Context &context) -> sys::ReturnCodes
 {
     std::string fileName = context.getBody()["fileName"].string_value();
+    auto path            = purefs::dir::os_updates / fileName;
+    auto fileExists      = vfs.fileExists(path.c_str());
+    if (fileExists) {
+        context.setResponseBody(json11::Json::object({{parserFSM::json::updateReady, true}}));
 
-    context.setResponseBody(json11::Json::object({{parserFSM::json::updateReady, true}}));
-
-    MessageHandler::putToSendQueue(context.createSimpleResponse());
-
-    auto msg = std::make_shared<sdesktop::UpdateOsMessage>(fileName, context.getUuid());
-    sys::Bus::SendUnicast(msg, service::name::service_desktop, ownerServicePtr);
-
-    return sys::ReturnCodes::Success;
+        auto msg = std::make_shared<sdesktop::UpdateOsMessage>(fileName, context.getUuid());
+        sys::Bus::SendUnicast(msg, service::name::service_desktop, ownerServicePtr);
+        MessageHandler::putToSendQueue(context.createSimpleResponse());
+        return sys::ReturnCodes::Success;
+    }
+    else {
+        context.setResponseBody(json11::Json::object({{parserFSM::json::updateReady, false}}));
+        context.setResponseStatus(http::Code::InternalServerError);
+        MessageHandler::putToSendQueue(context.createSimpleResponse());
+        return sys::ReturnCodes::Failure;
+    }
 }
 
 auto UpdateEndpoint::getUpdates(Context &context) -> sys::ReturnCodes